Chore: Move plugin settings store code into a dedicated service
