+++ pkg/services/pluginsettings/service/service.go
@@ -10,11 +10,12 @@ import (
 	"github.com/grafana/grafana/pkg/services/pluginsettings"
 	"github.com/grafana/grafana/pkg/services/secrets"
 	"github.com/grafana/grafana/pkg/services/sqlstore"
+	"github.com/grafana/grafana/pkg/services/sqlstore/db"
 )
 
-func ProvideService(store *sqlstore.SQLStore, secretsService secrets.Service) *Service {
+func ProvideService(db db.DB, secretsService secrets.Service) *Service {
 	s := &Service{
-		sqlStore: store,
+		db: db,
 		decryptionCache: secureJSONDecryptionCache{
 			cache: make(map[int64]cachedDecryptedJSON),
 		},
@@ -26,7 +27,7 @@ func ProvideService(store *sqlstore.SQLStore, secretsService secrets.Service) *S
 }
 
 type Service struct {
-	sqlStore        *sqlstore.SQLStore
+	db              db.DB
 	decryptionCache secureJSONDecryptionCache
 	secretsService  secrets.Service
 
@@ -44,7 +45,7 @@ type secureJSONDecryptionCache struct {
 }
 
 func (s *Service) GetPluginSettings(ctx context.Context, args *pluginsettings.GetArgs) ([]*pluginsettings.DTO, error) {
-	ps, err := s.sqlStore.GetPluginSettings(ctx, args.OrgID)
+	ps, err := s.getPluginSettings(ctx, args.OrgID)
 	if err != nil {
 		return nil, err
 	}
@@ -73,7 +74,7 @@ func (s *Service) GetPluginSettingByPluginID(ctx context.Context, args *pluginse
 		PluginId: args.PluginID,
 	}
 
-	err := s.sqlStore.GetPluginSettingById(ctx, query)
+	err := s.getPluginSettingById(ctx, query)
 	if err != nil {
 		return nil, err
 	}
@@ -97,7 +98,7 @@ func (s *Service) UpdatePluginSetting(ctx context.Context, args *pluginsettings.
 		return err
 	}
 
-	return s.sqlStore.UpdatePluginSetting(ctx, &models.UpdatePluginSettingCmd{
+	return s.updatePluginSetting(ctx, &models.UpdatePluginSettingCmd{
 		Enabled:                 args.Enabled,
 		Pinned:                  args.Pinned,
 		JsonData:                args.JSONData,
@@ -110,7 +111,7 @@ func (s *Service) UpdatePluginSetting(ctx context.Context, args *pluginsettings.
 }
 
 func (s *Service) UpdatePluginSettingPluginVersion(ctx context.Context, args *pluginsettings.UpdatePluginVersionArgs) error {
-	return s.sqlStore.UpdatePluginSettingVersion(ctx, &models.UpdatePluginSettingVersionCmd{
+	return s.updatePluginSettingVersion(ctx, &models.UpdatePluginSettingVersionCmd{
 		PluginVersion: args.PluginVersion,
 		PluginId:      args.PluginID,
 		OrgId:         args.OrgID,
@@ -137,4 +138,103 @@ func (s *Service) DecryptedValues(ps *pluginsettings.DTO) map[string]string {
 	}
 
 	return json
+}
+
+func (s *Service) getPluginSettings(ctx context.Context, orgID int64) ([]*models.PluginSetting, error) {
+	sql := `SELECT org_id, plugin_id, enabled, pinned, plugin_version FROM plugin_setting `
+	params := make([]interface{}, 0)
+
+	if orgID != 0 {
+		sql += "WHERE org_id=?"
+		params = append(params, orgID)
+	}
+
+	var rslt []*models.PluginSetting
+	err := s.db.WithDbSession(ctx, func(sess *sqlstore.DBSession) error {
+		return sess.SQL(sql, params...).Find(&rslt)
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return rslt, nil
+}
+
+func (s *Service) getPluginSettingById(ctx context.Context, query *models.GetPluginSettingByIdQuery) error {
+	return s.db.WithDbSession(ctx, func(sess *sqlstore.DBSession) error {
+		pluginSetting := models.PluginSetting{OrgId: query.OrgId, PluginId: query.PluginId}
+		has, err := sess.Get(&pluginSetting)
+		if err != nil {
+			return err
+		} else if !has {
+			return models.ErrPluginSettingNotFound
+		}
+		query.Result = &pluginSetting
+		return nil
+	})
+}
+
+func (s *Service) updatePluginSetting(ctx context.Context, cmd *models.UpdatePluginSettingCmd) error {
+	return s.db.WithTransactionalDbSession(ctx, func(sess *sqlstore.DBSession) error {
+		var pluginSetting models.PluginSetting
+
+		exists, err := sess.Where("org_id=? and plugin_id=?", cmd.OrgId, cmd.PluginId).Get(&pluginSetting)
+		if err != nil {
+			return err
+		}
+		sess.UseBool("enabled")
+		sess.UseBool("pinned")
+		if !exists {
+			pluginSetting = models.PluginSetting{
+				PluginId:       cmd.PluginId,
+				OrgId:          cmd.OrgId,
+				Enabled:        cmd.Enabled,
+				Pinned:         cmd.Pinned,
+				JsonData:       cmd.JsonData,
+				PluginVersion:  cmd.PluginVersion,
+				SecureJsonData: cmd.EncryptedSecureJsonData,
+				Created:        time.Now(),
+				Updated:        time.Now(),
+			}
+
+			// add state change event on commit success
+			sess.PublishAfterCommit(&models.PluginStateChangedEvent{
+				PluginId: cmd.PluginId,
+				OrgId:    cmd.OrgId,
+				Enabled:  cmd.Enabled,
+			})
+
+			_, err = sess.Insert(&pluginSetting)
+			return err
+		}
+
+		for key, encryptedData := range cmd.EncryptedSecureJsonData {
+			pluginSetting.SecureJsonData[key] = encryptedData
+		}
+
+		// add state change event on commit success
+		if pluginSetting.Enabled != cmd.Enabled {
+			sess.PublishAfterCommit(&models.PluginStateChangedEvent{
+				PluginId: cmd.PluginId,
+				OrgId:    cmd.OrgId,
+				Enabled:  cmd.Enabled,
+			})
+		}
+
+		pluginSetting.Updated = time.Now()
+		pluginSetting.Enabled = cmd.Enabled
+		pluginSetting.JsonData = cmd.JsonData
+		pluginSetting.Pinned = cmd.Pinned
+		pluginSetting.PluginVersion = cmd.PluginVersion
+
+		_, err = sess.ID(pluginSetting.Id).Update(&pluginSetting)
+		return err
+	})
+}
+
+func (s *Service) updatePluginSettingVersion(ctx context.Context, cmd *models.UpdatePluginSettingVersionCmd) error {
+	return s.db.WithTransactionalDbSession(ctx, func(sess *sqlstore.DBSession) error {
+		_, err := sess.Exec("UPDATE plugin_setting SET plugin_version=? WHERE org_id=? AND plugin_id=?", cmd.PluginVersion, cmd.OrgId, cmd.PluginId)
+		return err
+	})
 }
